rename files to match nice conventions
made of 5 commits:
TitleCase for components and their assets
move annotation view components into components/annoation
squash ArticleHeader into Article
use proper new filepaths
remove console.log
